Change default for retry-attempts to infinite retries
If not specified by the user, the default retry behaviour is to always
retry, with an exponential backoff wait time, up to the retry max wait
time (30 seconds by default).

If the user specify a retry-attempts, then the nubmer of retries will be
limited to that number.

A new env var is also introduced, 'retry-max-wait'. It is used to set
the max wait time in seconds between retries, with a default of 30
seconds.