Issue #533 - i18n.
Also replaced slf4j with JUL in the API module(as this needs to boil up to the spec, so we should not introduce new dependency)

For all other modules replaced slf4J with JBoss Logger and adding i18n.

I had to remove annotationprocessor in the compile plugin in the pom, but the build still works.

For some tests (like mosqito) I had to keep slf4j, but for most replace with JBoss Logger.
Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>