Fix rebalancing in the throttled commit strategy
Because of the execution model used by the Vert.x Kafka client, we may have started consuming records before the partitions assigned callback is invoked.
Also, when partitions are revoked, it must commit the processed offsets.  Unfortunately, with the current execution model, we cannot block the revocation, and the commit is a best-effort.