Fix duration unit mismatch with RabbitMQ reconnect interval
SM RabbitMQ’s units for `rabbitmq-reconnect-interval` (or `reconnect-interval` on a specific connector) is specified in _seconds_.    Vertx’s `NetworkOptions.reconnectInterval` is specified in milliseconds.

This leads to what an egregious hammering of the server when Vertx’s client is doing the reconnection because it’s trying every 10ms when using the defaults.