Allow declaring/binding to RabbitMQ server named queues
As detailed in the RabbitMQ docs, server named queues are important to work with `auto-delete` and `exclusive` queues without race conditions.

This allows naming the queue `(server.auto)` to create a server named queus. The `RabbitMQConnector` remaps this name to the empty string (which is what triggers the server named queue functionality).

Using a “remapping” of the queue name ensures that other code (e.g. logging) works and prints correctly.