Only allow a single ack or nack to be sent to RabbitMQ per message
Currently if the calls ack or nack more than a combine total of 1 time the client will send all the acks or nacks to the server. Since RabbitMQ explicitly disallows this it causes a connection error which then causes a complete connection restart.  The worste part is that without some real examination you cannot tell this is from multiply acking or nacking.

The best example of this is when a use `nacks` a message in a message handler that uses the `POST_PROCESSING` strategy. The framework will then `ack` or `nack` the message based on the outcome of the message handler.  No logging or hints exist to tell users that this is what is causing connection restarts.

This commit replaces the ack and nack handlers with versions that are no-ops; effectively ignoring any attempt to ack or nack after the first one.

Note that this will probably cause the message stream to fail as well; although it will do so with a relavent error message that a user can use to fix the code.