Replace ad-hoc conversion in IncomingRabbitMQMessage with converters
The ad-hoc conversion that was happening in `IncomingRabbitMQMessage.getPayload()` has been removed and replaced with a standard set of `IncomingMessageConverter`s that support JSON, text and byte arrays; the same that was attempted in the getPayload.

This ensures conversion only happens once (instead of every time getPayload is called) and that errors don’t happen or are handled properly when the do.

The removes `content-type-override` option. It was a hack to get the ad-hoc conversion to treat messages as some other type. This should be handled in an `IncomingMessageConverter` with proper error handling (which wasn’t done before.