Update to Vert.x 4.5.4
In this commit, an adjustment was made for MQTT functionality. Specifically, the hostname verifier now requires explicit configuration. Consequently, this commit introduces a new property to control the hostname verification (when connecting to the MQTT broker using TLS) and configures it with the `NONE` value (to avoid introducing a breaking change). You can toggle hostname verification by setting this property to "HTTPS".