Improve AMQP Client Shutdown Error Handling
Previously, when the AMQP client was closed and the broker was already stopped (such as in Quarkus tests), it resulted in a failure as the client couldn't disconnect properly. This failure was reported by the Mutiny dropped exception handler, potentially causing noise in the logs.

This commit introduces a safer approach by ignoring the failure during shutdown. Consequently, this failure will no longer be reported, addressing the potential noise in the logs. Importantly, this behavior only applies during application shutdown, ensuring that it doesn't affect normal operation.