Remove defunct properties (#3195)
In the new SmallRye parent, the `version.junit5` property has been discontinued. Using this property then leads to POM validation errors. However, it is not actually needed because the versions are managed in the parent, so just drop the usage of these properties.

This will allow the parent CI to include building this project, which should help avoid similar breakage in the future.